Enhance with :require_order option and string argument can start with '-'.
Add 'require_order' option so that first 'unknown' token terminates parsing
in the same way that '--' terminates parsing.  This option is manifested in
iterate_over_arguments function.

Fix process_desttype_arg so that an argument to a string-type option is NOT
processed to see if it looks like an option itself.  Allows the string argument
to start with a '-' character (allows passing of options as the argument to an
option).